telegram_stream_notify_bot/requirements.txt
aiogram==3.15.0
aiohttp==3.10.0
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.32.0
